renaming file for consistency
